refactor: moving setup to __tests__
